+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-07 13:59:17
+
+### Changed Files
+- `src/features/targets/TargetsPage.tsx`
+- `src/features/targets/components/PersistencyScenarios.tsx`
+- `tsconfig.tsbuildinfo`
+
+### Statistics
+```
+ 3 files changed, 293 insertions(+), 1 deletion(-)
+```
+
 ## [Checkpoint] - 2025-12-06 14:49:21
 
 ### Changed Files